fixes, helper methods, more font and text-path related tests

+++ trunk/raphael4gwt/src/org/sgx/raphael4gwt/raphael/base/Point.java
@@ -1,15 +1,23 @@
 package org.sgx.raphael4gwt.raphael.base;
 import com.google.gwt.core.client.JavaScriptObject;
-
+/**
+ *  TODO: if I dont use $wnd.Math.round() for ensure int gwt compielr complains in devmode.
+ * @author sg
+ *
+ */
 public class Point extends JavaScriptObject {
 	protected Point(){}
-	public native final int getX()/*-{
+	public native final double getX()/*-{
 		return this.x;
 	}-*/;
-	public native final int getY()/*-{
+	public native final double getY()/*-{
 		return this.y;
 	}-*/;
-	public native final int getAlpha()/*-{
+	/**
+	 * angle of derivative in degrees
+	 * @return
+	 */
+	public native final double getAlpha()/*-{
 		return this.alpha;
 	}-*/;
 	public native final String print()/*-{
